requirements.txt
copernicusmarine==2.3.0
python-dotenv==1.2.2
